WIP RFC: Sanitize paths also on Linux
I download most of my stuff on Linux machines but organize it on Windows machine. I find it most useful if the paths are already in a valid format so I can transfer the directories directly to my Windows partitions.
I would not mind if there is a command line option (like --sanitizepaths) to make the Linux one optional but this is beyond my capabilities.