Fix interruped downloads with --latest-stamps (#1206)
The timestamp of the most recent post is cached in NodeIterator (and
saved to the disk), and this timestamp is stored, instead of the
timestamp instaloader was run.

This way, even in later resuming runs the timestamp stored is the same
that would have been stored if the first run.