Added functions `is_reposted` and `original_poster_profile` under `StoryItem` class
A lot of times the "owner" is reposting a story he was tagged into. I am calling the person who tagged the "owner" as "original_poster".

The GraphQL query from such a reposted story has information inside the `tappable_objects` key of each `StoryItem`.
For example for a reposted story it looks like this:

```
"tappable_objects": [
                        {
                            "__typename": "GraphTappableMention",
                            "x": 0.49967163378806606,
                            "y": 0.5150924938267181,
                            "width": 0.9216036185381431,
                            "height": 0.9216036185381431,
                            "rotation": 0.0,
                            "custom_title": null,
                            "attribution": null,
                            "username": "username",
                            "full_name": "User Name",
                            "is_private": false
                        }
                    ]
```

And for a story that was NOT reposted there is nothing inside `tappable_objects` with the `__typename` of `GraphTappableMention`.

The functions I added are pretty self explanatory and I followed the structure the rest of the project has.